Add unit tests for Logger and fix compiler warnings
- Create tests/test_Logger.cpp to test Logger functionality:
  - Verify write buffering and accumulation
  - Verify circular buffer behavior (overwriting old logs)
  - Verify log level filtering
- Update tests/Makefile to include test_Logger target
- Update tests/mocks/Arduino.h:
  - Add buffering to MockSerial for verification
  - Add indexOf to MockString
- Update tests/mocks/freertos/FreeRTOS.h:
  - Add pdTRUE/pdFALSE definitions
- Update tests/mocks/mocks.cpp:
  - Conditionally compile Logger mocks to allow linking real Logger
- Fix signed/unsigned comparison warnings in src/Logger.cpp

+++ src/Logger.cpp
@@ -43,7 +43,7 @@ size_t Logger::printf(const char *format, ...) {
     va_end(arg);
     return 0;
   };
-  if (len >= sizeof(loc_buf)) {
+  if ((size_t)len >= sizeof(loc_buf)) {
     temp = (char *)malloc(len + 1);
     if (temp == NULL) {
       va_end(arg);
@@ -78,7 +78,7 @@ void Logger::debug(const char *format, ...) {
     va_end(arg);
     return;
   };
-  if (len >= sizeof(loc_buf)) {
+  if ((size_t)len >= sizeof(loc_buf)) {
     temp = (char *)malloc(len + 1);
     if (temp == NULL) {
       va_end(arg);
